10.6.1 전체 테스트

+++ fastapi/src/data/creature.py
@@ -1,5 +1,7 @@
 from data.init_db import conn, curs
 from model.creature import Creature
+from error import Missing, Duplicate
+from sqlite3 import IntegrityError
 
 curs.execute("""create table if not exists creature(
                     name text primary key, 
@@ -20,7 +22,10 @@ def get_one(name: str) -> Creature:
     params = {"name": name}
     curs.execute(qry, params)
     row = curs.fetchone()
-    return row_to_model(row)
+    if row:
+        return row_to_model(row)
+    else:
+        raise Missing(msg=f"Creature {name} not found")
 
 def get_all() -> list[Creature]:
     qry = "select * from creature"
@@ -29,44 +34,51 @@ def get_all() -> list[Creature]:
     return [row_to_model(row) for row in rows]
 
 def create(creature: Creature):
+    if not creature:
+        return None
     qry = """insert into creature values
         (:name, :description, :country, :area, :aka)"""
     params = model_to_dict(creature)
-    curs.execute(qry, params)
-    conn.commit()
+    try:
+        curs.execute(qry, params)
+        conn.commit()
+    except IntegrityError:
+        raise Duplicate(msg=
+                        f"Creature {creature.name} already exists")
     return get_one(creature.name)
+
+def update(name: str, creature: Creature):
+    if not (name and creature):
+        return None
+    qry = """update creature
+                set country=:country,
+                    name=:name,
+                    description=:description,
+                    area=:area,
+                    aka=:aka
+                where name=:name_orig"""
+    params = model_to_dict(creature)
+    params["name_orig"] = name
+    _ = curs.execute(qry, params)
+    conn.commit()
+    if curs.rowcount == 1:
+        return get_one(creature.name)
+    else:
+        raise Missing(msg=f"Creature {name} not fount")
 
 def modify(name: str, creature: Creature):
-    qry = """update creature
-                set country=:country,
-                    name=:name,
-                    description=:description,
-                    area=:area,
-                    aka=:aka
-                where name=:name_orig"""
-    params = model_to_dict(creature)
-    params["name_orig"] = name
-    _ = curs.execute(qry, params)
-    conn.commit()
-    return get_one(creature.name)
+    return update(name, creature)
 
 def replace(name: str, creature: Creature):
-    qry = """update creature
-                set country=:country,
-                    name=:name,
-                    description=:description,
-                    area=:area,
-                    aka=:aka
-                where name=:name_orig"""
-    params = model_to_dict(creature)
-    params["name_orig"] = name
-    _ = curs.execute(qry, params)
-    conn.commit()
-    return get_one(creature.name)
+    return update(name, creature)
 
 def delete(name: str) -> bool:
+    if not name:
+        return False
     qry = "delete from creature where name = :name"
     params = {"name": name}
     res = curs.execute(qry, params)
     conn.commit()
+    if curs.rowcount != 1:
+        raise Missing(msg=f"Creature {name} not found")
     return bool(res)
